Move tests to core
