+++ src/Controls/tests/DeviceTests/Elements/VisualElementTests.cs
@@ -165,54 +165,5 @@ namespace Microsoft.Maui.DeviceTests
 				Assert.Equal(1, unloaded);
 			});
 		}
-
-		string TooltipTextForHandler(ViewHandler handler)
-		{
-#if IOS
-			return handler.ToPlatform()?.GetToolTipInteraction()?.DefaultToolTip;
-#elif ANDROID
-				return handler.ToPlatform()?.TooltipText;
-#elif WINDOWS
-				return Microsoft.UI.Xaml.Controls.ToolTipService.GetToolTip(handler.ToPlatform()) as string;
-#endif
-		}
-
-		const string ToolTipTextStringValue = "ToolTip Text Here!";
-
-		[Theory]
-		[InlineData(ToolTipTextStringValue)]
-		[InlineData(null)]
-		public Task ToolTipTextAppliedToLabel(string text)
-			=> AssertTooltipTextApplied<Label, LabelHandler>(text);
-
-		[Theory]
-		[InlineData(ToolTipTextStringValue)]
-		[InlineData(null)]
-		public Task ToolTipTextAppliedToButton(string text)
-			=> AssertTooltipTextApplied<Button, ButtonHandler>(text);
-
-		[Theory]
-		[InlineData(ToolTipTextStringValue)]
-		[InlineData(null)]
-		public Task ToolTipTextAppliedToImage(string text)
-			=> AssertTooltipTextApplied<Image, ImageHandler>(text);
-
-		[Theory]
-		[InlineData(ToolTipTextStringValue)]
-		[InlineData(null)]
-		public Task ToolTipTextAppliedToCheckbox(string text)
-			=> AssertTooltipTextApplied<CheckBox, CheckBoxHandler>(text);
-
-
-		async Task AssertTooltipTextApplied<TVisualElement, THandler>(string expected)
-			where THandler : ViewHandler
-			where TVisualElement : VisualElement
-		{
-			var control = (TVisualElement)Activator.CreateInstance(typeof(TVisualElement));
-			var handler = await CreateHandlerAsync<THandler>(control);
-			await InvokeOnMainThreadAsync(() => control.TooltipText = expected);
-			var platformText = await InvokeOnMainThreadAsync(() => TooltipTextForHandler(handler));
-			Assert.Equal(expected, platformText);
-		}
 	}
 }
