+++ src/Core/tests/DeviceTests/Handlers/View/ViewHandlerTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Microsoft.Maui;
 using Microsoft.Maui.DeviceTests.Stubs;
 using Microsoft.Maui.Graphics;
@@ -55,6 +56,56 @@ namespace Microsoft.Maui.DeviceTests
 			});
 
 			Assert.Equal(2, didUpdateFrame);
+		}
+
+		const string ToolTipTextStringValue = "ToolTip Text Here!";
+
+		[Theory]
+		[InlineData(ToolTipTextStringValue)]
+		[InlineData(null)]
+		public Task ToolTipTextAppliedToLabel(string text)
+			=> AssertTooltipTextApplied<LabelStub>(text);
+
+		[Theory]
+		[InlineData(ToolTipTextStringValue)]
+		[InlineData(null)]
+		public Task ToolTipTextAppliedToButton(string text)
+			=> AssertTooltipTextApplied<ButtonStub>(text);
+
+		[Theory]
+		[InlineData(ToolTipTextStringValue)]
+		[InlineData(null)]
+		public Task ToolTipTextAppliedToImage(string text)
+			=> AssertTooltipTextApplied<ImageStub>(text);
+
+		[Theory]
+		[InlineData(ToolTipTextStringValue)]
+		[InlineData(null)]
+		public Task ToolTipTextAppliedToCheckbox(string text)
+			=> AssertTooltipTextApplied<CheckBoxStub>(text);
+
+		async Task AssertTooltipTextApplied<TElement>(string expected)
+			where TElement : StubBase
+		{
+			var control = (TElement)Activator.CreateInstance(typeof(TElement));
+			control.TooltipText = expected;
+
+			var handler = await CreateHandlerAsync(control);
+
+			var platformText = await InvokeOnMainThreadAsync(() =>
+			{
+#if IOS || MACCATALYST
+				return handler.ToPlatform()?.GetToolTipInteraction()?.DefaultToolTip;
+#elif ANDROID
+				return handler.ToPlatform()?.TooltipText;
+#elif WINDOWS
+				return Microsoft.UI.Xaml.Controls.ToolTipService.GetToolTip(handler.ToPlatform()) as string;
+#else
+				return null;
+#endif
+			});
+
+			Assert.Equal(expected, platformText);
 		}
 	}
 }
